Download files in proper threads

+++ android/app/src/main/java/com/receiptsmobile/files/DownloadResult.java
@@ -2,16 +2,16 @@ package com.receiptsmobile.files;
 
 public class DownloadResult {
     public final int statusCode;
-    public final int bytesWritten;
+    public final long bytesWritten;
     public final Exception exception;
 
-    public DownloadResult(int statusCode, int bytesWritten, Exception exception) {
+    public DownloadResult(int statusCode, long bytesWritten, Exception exception) {
         this.statusCode = statusCode;
         this.bytesWritten = bytesWritten;
         this.exception = exception;
     }
 
-    public static DownloadResult ok(int statusCode, int bytesWritten) {
+    public static DownloadResult ok(int statusCode, long bytesWritten) {
         return new DownloadResult(statusCode, bytesWritten, null);
     }
